rename files AlleleVariantsSummary and VariantSummary
