+++ src/containers/allelePage/VariantSummary.js
